requirements.txt
torch
torchvision
scipy
numpy
pandas
pyyaml
seaborn
datetime
pyyaml

